chore: refactored code
